Use %lld in printf for 64-bit time_t
musl 1.2.0 switched to use 64-bit time_t everywhere, including 32-bit
architectures, causing garbage values to be printed from ppswatch and
ppstest.

Use the correct format string for the affected printf statements and
explicitly cast to long long to avoid potential compatibility issues
with 32-bit glibc.

+++ ppstest.c
@@ -110,13 +110,13 @@ retry:
 	}
 
 	printf("source %d - "
-	       "assert %ld.%09ld, sequence: %ld - "
-	       "clear  %ld.%09ld, sequence: %ld\n",
+	       "assert %lld.%09ld, sequence: %ld - "
+	       "clear  %lld.%09ld, sequence: %ld\n",
 	       i,
-	       infobuf.assert_timestamp.tv_sec,
+	       (long long)infobuf.assert_timestamp.tv_sec,
 	       infobuf.assert_timestamp.tv_nsec,
 	       infobuf.assert_sequence,
-	       infobuf.clear_timestamp.tv_sec,
+	       (long long)infobuf.clear_timestamp.tv_sec,
 	       infobuf.clear_timestamp.tv_nsec, infobuf.clear_sequence);
 	fflush(stdout);
